Preserve terminal interactions across capped gutters
